+++ libpod/driver/driver.go
@@ -1,12 +1,9 @@
 package driver
 
-import cstorage "github.com/containers/storage"
-
-// Data handles the data for a storage driver
-type Data struct {
-	Name string            `json:"Name"`
-	Data map[string]string `json:"Data"`
-}
+import (
+	cstorage "github.com/containers/storage"
+	"github.com/projectatomic/libpod/pkg/inspect"
+)
 
 // GetDriverName returns the name of the driver for the given store
 func GetDriverName(store cstorage.Store) (string, error) {
@@ -27,7 +24,7 @@ func GetDriverMetadata(store cstorage.Store, layerID string) (map[string]string,
 }
 
 // GetDriverData returns the Data struct with information of the driver used by the store
-func GetDriverData(store cstorage.Store, layerID string) (*Data, error) {
+func GetDriverData(store cstorage.Store, layerID string) (*inspect.Data, error) {
 	name, err := GetDriverName(store)
 	if err != nil {
 		return nil, err
@@ -36,7 +33,7 @@ func GetDriverData(store cstorage.Store, layerID string) (*Data, error) {
 	if err != nil {
 		return nil, err
 	}
-	return &Data{
+	return &inspect.Data{
 		Name: name,
 		Data: metaData,
 	}, nil
